feat(shortcuts): extract openOptionsWithHash + add cross-browser shortcuts URL helper
Move the inline openOptionsWithHash helper out of PopupToolbar into
src/utils/openOptions.ts so it can be reused from the background command
handler and any future caller.

Add src/utils/browserUrls.ts with getShortcutsCustomizeInfo() that returns
the right customization URL per browser (chrome://extensions/shortcuts on
Chromium, about:addons on Firefox where extensions cannot link directly to
the shortcuts panel) along with an isDirect flag callers can use to display
a manual hint when needed.
